requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
pydantic==2.11.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.32.5
